put the queue in its own module
